requirements.txt
Flask
google-cloud-secret-manager
google-generativeai
fastmcp
requests
